Day 5 part 1

+++ AdventOfCode.2023.Tdd.Xunit/Week1/Day5/Domain/MapSources/Fertilizer.cs
@@ -2,5 +2,5 @@
 
 public class Fertilizer
 {
-    public int Id { get; set; }
+    public double Id { get; set; }
 }
